build: Split frontend apps and update docker config
The mounting of node_modules in the containers was causing issues,
so the volume mounting has been moved to compose.frontend.yml. The
pnpm workspace was also removed as it was not being used.

eslint was also downgraded to work with the old config format for
now until some other dependent packages are updated.